requirements.txt
Flask==2.3.3
mysql-connector-python==8.2.0
Werkzeug==2.3.7
PyPDF2==3.0.1
python-dotenv==0.19.2
email-validator==2.1.0.post1
Flask-WTF==1.0.0
pytest==7.0.0
pytest-flask==1.2.0
google-generativeai==0.3.1
PyMuPDF==1.23.7
requests==2.28.1
protobuf==3.20.3
google-api-core==2.11.0
google-auth==2.17.3
MarkupSafe==2.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
click==8.1.7
blinker==1.6.3 